Refactored code ready for unit tests (BREAKING CHANGES!)
* Events are now handled in WebsocketListener
* StreamDeck4J is now Plugin (interface)
* There is an implementation for easy access to "internal" methods now through PluginImpl
* Events now all have a `getPlugin()` function to return Plugin
* Updated SD4JMain to show these changes
