+++ src/main/java/com/walshydev/streamdeck4j/SD4JMain.java
@@ -17,35 +17,34 @@ public class SD4JMain {
 
     public static void main(String[] args) {
         logger.info("Starting main test app for StreamDeck4J");
-        StreamDeck4J streamDeck4J = new StreamDeck4J();
-
-        streamDeck4J.addListener(new AbstractListener() {
-            @Override
-            public void onActionAppeared(ActionAppearedEvent event) {
-                streamDeck4J.setTitle(
-                    event.getContext(),
-                    "Testing123\n" + System.currentTimeMillis(),
-                    Destination.HARDWARE_AND_SOFTWARE
-                );
-            }
-
-            @Override
-            public void onKeyDown(KeyDownEvent event) {
-                try {
-                    streamDeck4J.openURL(new URL("https://github.com/WalshyDev/StreamDeck4J"));
-                    Thread.sleep(2000);
-                    streamDeck4J.setImage(
+        Plugin plugin = new PluginBuilder(args)
+            .registerListener(new AbstractListener() {
+                @Override
+                public void onActionAppeared(ActionAppearedEvent event) {
+                    event.getPlugin().setTitle(
                         event.getContext(),
-                        ImageIO.read(getClass().getClassLoader().getResourceAsStream("blob.png")),
+                        "Testing123\n" + System.currentTimeMillis(),
                         Destination.HARDWARE_AND_SOFTWARE
                     );
-                } catch (InterruptedException | IOException e) {
-                    e.printStackTrace();
                 }
-            }
-        });
 
-        logger.info("Connecting!");
-        streamDeck4J.connect(args);
+                @Override
+                public void onKeyDown(KeyDownEvent event) {
+                    try {
+                        event.getPlugin().openURL(new URL("https://github.com/WalshyDev/StreamDeck4J"));
+                        Thread.sleep(2000);
+                        event.getPlugin().setImage(
+                            event.getContext(),
+                            ImageIO.read(getClass().getClassLoader().getResourceAsStream("blob.png")),
+                            Destination.HARDWARE_AND_SOFTWARE
+                        );
+                    } catch (InterruptedException | IOException e) {
+                        e.printStackTrace();
+                    }
+                }
+            })
+            .buildAsync();
+
+        logger.info("Using StreamDeck version: {}", plugin.getApplication().getVersion());
     }
 }
