+++ src/main/java/com/walshydev/streamdeck4j/events/ActionDisappearedEvent.java
@@ -1,6 +1,7 @@
 package com.walshydev.streamdeck4j.events;
 
 import com.google.gson.JsonObject;
+import com.walshydev.streamdeck4j.Plugin;
 import com.walshydev.streamdeck4j.info.Coordinates;
 import lombok.Getter;
 
@@ -17,6 +18,7 @@ public class ActionDisappearedEvent extends Event {
     private final boolean multiAction;
 
     public ActionDisappearedEvent(
+        @Nonnull Plugin plugin,
         @Nonnull String context,
         @Nonnull String action,
         @Nonnull String deviceId,
@@ -25,7 +27,7 @@ public class ActionDisappearedEvent extends Event {
         int state,
         boolean multiAction
     ) {
-        super(context);
+        super(plugin, context);
         this.action = action;
         this.deviceId = deviceId;
         this.settings = settings;
