+++ src/main/java/com/walshydev/streamdeck4j/events/DeviceDisconnectedEvent.java
@@ -1,5 +1,6 @@
 package com.walshydev.streamdeck4j.events;
 
+import com.walshydev.streamdeck4j.Plugin;
 import lombok.Getter;
 
 import javax.annotation.Nonnull;
@@ -10,9 +11,10 @@ public class DeviceDisconnectedEvent extends Event {
     private final String deviceId;
 
     public DeviceDisconnectedEvent(
+        @Nonnull Plugin plugin,
         @Nonnull String deviceId
     ) {
-        super(null);
+        super(plugin, null);
         this.deviceId = deviceId;
     }
