+++ src/main/java/com/walshydev/streamdeck4j/events/SentToPluginEvent.java
@@ -1,6 +1,7 @@
 package com.walshydev.streamdeck4j.events;
 
 import com.google.gson.JsonObject;
+import com.walshydev.streamdeck4j.Plugin;
 import lombok.Getter;
 
 import javax.annotation.Nonnull;
@@ -12,11 +13,12 @@ public class SentToPluginEvent extends Event {
     private final JsonObject payload;
 
     public SentToPluginEvent(
+        @Nonnull Plugin plugin,
         @Nonnull String context,
         @Nonnull String action,
         @Nonnull JsonObject payload
     ) {
-        super(context);
+        super(plugin, context);
         this.action = action;
         this.payload = payload;
     }
